formatting and actions updates

+++ src/routes/api/user/[userId]/+server.ts
@@ -3,19 +3,24 @@ import { error, json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 
 export const DELETE: RequestHandler = async ({ params, locals }) => {
-	const userData = (await adminDB.collection('users').doc(locals.userID!).get()).data();
+	if (!locals.userID) {
+		throw error(401, 'You must be logged in to do this.');
+	}
+
+	const userData = (await adminDB.collection('users').doc(locals.userID).get()).data();
 
 	if (!userData || !userData.permissions || userData.permissions !== 'admin') {
 		throw error(401, 'You must be an admin to do this.');
 	}
 
-	await adminAuth.deleteUser(params.userId)
+	await adminAuth
+		.deleteUser(params.userId)
 		.then(() => {
 			adminDB.collection('users').doc(params.userId).delete();
 		})
 		.catch((err) => {
-            console.log(err.message);
+			console.log(err.message);
 			throw error(500, err);
 		});
-    return json({status: 'User Deleted'});
+	return json({ status: 'User Deleted' });
 };
